# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Freightliner Container Registry Replication API
  description: |
    RESTful API for container image replication across multiple registries.
    Supports AWS ECR, GCP GCR, Azure ACR, Docker Hub, Harbor, and generic OCI registries.
  version: 1.0.0
  contact:
    name: Freightliner Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://freightliner.example.com/api/v1
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      description: Returns a list of replication jobs
      operationId: listJobs
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [replicate, replicate-tree, checkpoint]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /jobs/{id}/cancel:
    post:
      tags: [Jobs]
      summary: Cancel a job
      description: Cancels a running or pending job
      operationId: cancelJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /checkpoints:
    get:
      tags: [Checkpoints]
      summary: List checkpoints
      description: Returns all available checkpoints for tree replication
      operationId: listCheckpoints
      responses:
        '200':
          description: List of checkpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkpoints:
                    type: array
                    items:
                      $ref: '#/components/schemas/Checkpoint'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /checkpoints/{id}:
    get:
      tags: [Checkpoints]
      summary: Get checkpoint details
      description: Returns detailed information about a specific checkpoint
      operationId: getCheckpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Checkpoint details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checkpoint'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Checkpoints]
      summary: Delete a checkpoint
      description: Removes a checkpoint from storage
      operationId: deleteCheckpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Checkpoint deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [replicate, replicate-tree, checkpoint]
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        source:
          type: string
        destination:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration:
          type: number
          description: Duration in seconds
        progress:
          type: object
          properties:
            current:
              type: integer
            total:
              type: integer
            percentage:
              type: number
        result:
          type: object
          description: Job-specific result data
        error:
          type: string
          description: Error message if job failed
...
    Checkpoint:
      type: object
      properties:
        id:
          type: string
        job_id:
          type: string
        created_at:
          type: string
          format: date-time
        source:
          type: string
        destination:
          type: string
        progress:
          type: object
        completed_repos:
          type: array
          items:
            type: string
        failed_repos:
          type: array
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: string
            format: date-time
          description: Time when rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...